# Better performance (especially on large files).