--- a/.pre-commit-config.yaml
+++ b/.pre-commit-config.yaml
@@ -1,0 +1,5 @@
+repos:
+- repo: https://github.com/godaddy/tartufo
+  rev: v2.6.0
+  hooks:
+  - id: tartufo