--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -1,6 +1,6 @@
-builder:
-  image: starefossen/iojs-imagemagick:1.6-6.9
+dev:
+  image: starefossen/iojs-imagemagick:1-6
   working_dir: /usr/src/app
   volumes:
     - ".:/usr/src/app"
-  command: "node polling.js"
+  command: "npm run watch"