--- a/index.js
+++ b/index.js
@@ -24,7 +24,7 @@ module.exports.cmd = function(path, opts) {
   opts = opts || {};
   var format = [
     'name=',
-    'size=%[size]',
+    'size=%b',
     'format=%m',
     'colorspace=%[colorspace]',
     'height=%[height]',
@@ -33,7 +33,7 @@ module.exports.cmd = function(path, opts) {
     (opts.exif ? '%[exif:*]' : '')
   ].join("\n");
 
-  return 'identify -format "' + format + '" ' + path;
+  return 'convert -ping ' + path + ' -format "' + format + '" info:';
 };
 
 module.exports.parse = function(path, stdout, opts) {