--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "im-metadata",
-  "version": "1.0.2",
+  "version": "1.0.3",
   "description": "Retrieve image metadata using ImageMagick's identify command",
   "main": "index.js",
   "scripts": {