--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "im-metadata",
-  "version": "1.0.3",
+  "version": "1.1.0",
   "description": "Retrieve image metadata using ImageMagick's identify command",
   "main": "index.js",
   "scripts": {