--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "im-metadata",
-  "version": "2.1.0",
+  "version": "2.1.1",
   "description": "Retrieve image metadata using ImageMagick's identify command",
   "main": "index.js",
   "scripts": {