--- a/package.json
+++ b/package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "mocha test.js",
+    "hint": "./node_modules/.bin/jshint index.js test.js",
     "watch": "./node_modules/.bin/mocha -w test.js"
   },
   "repository": {
@@ -25,6 +26,7 @@
   },
   "homepage": "https://github.com/Turistforeningen/node-im-metadata",
   "devDependencies": {
+    "jshint": "^2.8.0",
     "mocha": "^2.2.4"
   }
 }