--- a/test.js
+++ b/test.js
@@ -109,12 +109,7 @@ describe('metadata()', function() {
       assert.equal(data.colorspace, 'RGB');
       assert.equal(data.height, 3456);
       assert.equal(data.width, 5184);
-
-      // Ok, I give up. For some reason there is this inconsistency between
-      // ImageMagick identify versions which yilds undefined/empty string for
-      // the orientation on the CI server. I can not find any reference on this
-      // issue which is why this test will not be run on the CI server.
-      if (!process.env.CI) { assert.equal(data.orientation, 'TopLeft'); }
+      assert.equal(data.orientation, 'TopLeft');
 
       assert.equal(typeof data.exif, 'undefined');
 
@@ -133,12 +128,7 @@ describe('metadata()', function() {
       assert.equal(data.colorspace, 'RGB');
       assert.equal(data.height, 3456);
       assert.equal(data.width, 5184);
-
-      // Ok, I give up. For some reason there is this inconsistency between
-      // ImageMagick identify versions which yilds undefined/empty string for
-      // the orientation on the CI server. I can not find any reference on this
-      // issue which is why this test will not be run on the CI server.
-      if (!process.env.CI) { assert.equal(data.orientation, 'TopLeft'); }
+      assert.equal(data.orientation, 'TopLeft');
 
       assert.equal(typeof data.exif, 'object');
       assert.equal(Object.keys(data.exif).length, 36);