--- a/test.js
+++ b/test.js
@@ -5,17 +5,17 @@ var metadata = require('./index');
 
 describe('metadata.cmd()', function() {
   it('returns command without exif data', function() {
-    var cmd = 'identify -format "name=\nsize=%[size]\nformat=%m\n'
+    var cmd = 'convert -ping /foo/bar/baz -format "name=\nsize=%b\nformat=%m\n'
             + 'colorspace=%[colorspace]\nheight=%[height]\nwidth=%[width]\n'
-            + 'orientation=%[orientation]\n" /foo/bar/baz';
+            + 'orientation=%[orientation]\n" info:';
 
     assert.equal(metadata.cmd('/foo/bar/baz'), cmd);
   });
 
   it('returns command with exif data', function() {
-    var cmd = 'identify -format "name=\nsize=%[size]\nformat=%m\n'
+    var cmd = 'convert -ping /foo/bar/baz -format "name=\nsize=%b\nformat=%m\n'
             + 'colorspace=%[colorspace]\nheight=%[height]\nwidth=%[width]\n'
-            + 'orientation=%[orientation]\n%[exif:*]" /foo/bar/baz';
+            + 'orientation=%[orientation]\n%[exif:*]" info:';
 
     assert.equal(metadata.cmd('/foo/bar/baz', {exif: true}), cmd);
   });
@@ -111,7 +111,9 @@ describe('metadata()', function() {
 
       assert.equal(data.path, './assets/image.jpg');
       assert.equal(data.name, '');
-      assert.equal(data.size, 4504682);
+      // allow for precision errors. image magick returns 5 decimal places for
+      // me locally on OSX vs 3 decimal places on linux.
+      assert.equal(data.size - data.size % 1000, 4504000);
       assert.equal(data.format, 'JPEG');
       assert.equal(data.colorspace, 'RGB');
       assert.equal(data.height, 3456);
@@ -130,7 +132,7 @@ describe('metadata()', function() {
 
       assert.equal(data.path, './assets/image.jpg');
       assert.equal(data.name, '');
-      assert.equal(data.size, 4504682);
+      assert.equal(data.size - data.size % 1000, 4504000);
       assert.equal(data.format, 'JPEG');
       assert.equal(data.colorspace, 'RGB');
       assert.equal(data.height, 3456);