Dealing with large images (especially SVG images with a large viewport) can take a long time and use an excessive amount of memory or even die with a resources exhausted error.

Using this test file:
```svg
<?xml version="1.0" encoding="UTF-8"?>
<svg version="1.2" viewBox="0 0 20000 20000">
  <circle cx="10000" cy="10000" r="8000" stroke="black" stroke-width="30" fill="red" />
</svg>
```

Using the `time` function and the `memusg` function found here: https://gist.github.com/netj/526585

```
identify -format "%m %[size] %[width] %[height] %[colorspace]" test.svg
time: 62.56s user 8.36s system 92% cpu 1:16.41 total
memusg: peak=7578552
```

This alternative implementation uses ImageMagick's `convert -ping` function which doesn't keep any pixel information.

```
convert -ping test.svg -format "%m %[size] %[width] %[height] %[colorspace]" info:
time: 0.01s user 0.01s system 51% cpu 0.038 total
memusg: peak=940
```

Significantly better.

I've also added the `-precision 12` option. There were some failing tests on my local OSX box since it seems the default precision for OSX is 6 while Linux is 4. Setting it to 12 removes that difference and also gives better accuracy.